+++ slices/user.ts
@@ -2,7 +2,6 @@ import { createSlice } from "@reduxjs/toolkit";
 import { HYDRATE } from "next-redux-wrapper";
 
 const initialState = {
-  signinLoading: false,
   email: "",
   nickname: "",
   profileImage: null,
@@ -13,21 +12,8 @@ const userSlice = createSlice({
   name: "user",
   initialState,
   reducers: {
-    signin(state, action) {
-      state.signinLoading = true;
-      state.email = "";
-      state.nickname = "";
-      state.profileImage = null;
-      state.accessToken = "";
-    },
     setAccessToken(state, action) {
       state.accessToken = action.payload.accessToken;
-    },
-    signinSuccess(state) {
-      state.signinLoading = false;
-    },
-    signinFailure(state) {
-      state.signinLoading = false;
     },
   },
   extraReducers: {
@@ -40,7 +26,6 @@ const userSlice = createSlice({
   },
 });
 
-export const { signin, setAccessToken, signinSuccess, signinFailure } =
-  userSlice.actions;
+export const { setAccessToken } = userSlice.actions;
 
 export default userSlice;
